Fix problems with IAnimal.isInLove()
1) The IAnimal interface defines the isInLove() method but none of the animals
which implement the interface were providing an implementation for it.
Since they all also derive from EntityAnimal, they were picking up its
isInLove() method, which won't work. So I added isInLove() methods to all
of the classes which implement IAnimal.

2) Since all of the animals which implement IAnimal also derive from
EntityAnimal, IAnimal.isInLove() was shadowing EntityAnimal.isInLove().
This resulting in MCP obfuscating the former (a bug in MCP?). However the
callers of IAnimal.isInLove() were not being similarly obfuscated, leading
to AbstractMethod exceptions in the obfuscated binary.

To fix that I changed the name of IAnimal.isInLove() to
IAnimal.getInLove().