Fix bugs in release scripts.
Corrected the case on several filenames to match their case in git, so that
the scripts will work on case-sensitive file systems.

In release.py main() was creating 'forge/mcp/TFCraft/TFC Build' while
zip_start was expecting 'forge/mcp/TFC Build'. Fixed the former to use the
same path as the latter.