Added Vinegar and Cheese items; along with corresponding recipes.
Vinegar is made by re-sealing a barrel of alcohol.
Cheese is made by adding milk to a barrel and at least 1 bucket of vinegar and then sealing the barrel

Male sheep now have horns

Added recipe to make white carpet from wool cloth, and recipes to dye carpets

Added feature to the "empty" function of the barrel which will fill an empty bucket in the barrel's slot instead of emptying the barrel. Note: this only works for valid liquids, ie: water, limewater, vinegar and milk.