Added QuernManager to the API
Quern recipies can now be added by other mods.
Quern input slot takes only input items that are used in a registered
recipe.
If an output is different from what is being processed it will be
ejected to make room for the new item, if it's the same, it will be
added to the output if possible. If the stack reaches it's max size, it
will be ejected and what is left over will be added to the output slot.