Added crop infestations that have a 1 in 2000 chance of occurring for each crop. When an infestation occurs, the entire chunk is considered infested and depending on the level of infestation, the farmland block itself gets infested. Just removing these blocks wont really help because the chunk itself is considered infested as well and the infestation will take hold again. Connected farmland will spread the infestation into blocks next to them. Infestation spread this way will also infect the chunk it enters if that chunk did not have infestation.
The effects of an infested crop is a halving of crop growth rate.

The only way to remove or prevent infestation would be to either wait until it gets cold enough for the bugs to die or plant certain flowers near your gardens which ward off bad bugs(TBD).

The current infestation texture needs an update as it is programmer art.