Fix AnvilRecipe returning mutable result.
Before this fix, with the output of a weld in WELDOUT_SLOT and
the same item type in INPUT1_SLOT, picking up the item in WELDOUT
and moving to INPUT1 swapped the two ItemStacks. Applying any anvil
work to the now INPUT1 item stack was actually modifying the
AnvilRecipe.result.