This fixes the pitkiln instant cooking.
If there is a fire on top of, or around the pitkiln, the pitkiln will
ignite and start the burn countdown correctly without instantly cooking
the clay items. If the pitkiln is invalid(a state between 1 straw and 7
wood), all items will get ejected and the TE will be removed.