Fixed spawn protection
spawnProtectionTimer was not being updated back to the nbt correctly,
which added "TFCOptions.protectionGain" to the protection on every
player tick maxing it out in a few sec.
Also fixed the catching-up-to-time-loop in spawn meter TE.
If the meter starts with 0, depending on the server time,
it will need a lot of loops adding 1000 to get to the right time,
now it will start from the block placement time.
This fixes a possible lag issue.