aforementioned fish code was actually there all along.
Added horse health bars when riding.

Introduced mechanic where animals can become familiar with players over time. Currently, wolves have this feature alone, but support for all animals is planned. When a bone is used to "tame" a wolf, it will follow the player who tamed it. It cannot be told to sit and will still attempt to kill wild animals. It won't have a collar and won't protect the player from monsters. You can increase the familiarity by performing a particular type of interaction with the animal at least once per day. For wolves, this means shift-right clicking. The adult wolf tamed with a bone from the wild can never actually be tamed, but baby wolves can have their familiarity increased a lot faster and higher than an adult. If this value surpasses "80", the wolf will be tamed.