Added a valid tool list that can be used for chopping down trees.
Added a logCount variable to the schematics class which is being used to
determine if a tool is able to chop a tree or if its too damaged. This
also doesn't allow a stone axe to cut a tree that would damage it above
it's max damage value. The natural log block can not be broken without
a valid tool, except for explosions and fire which still needs to be
fixed. Also added tool break sound and animation.