Smoke Rack GUI Interface.
1. Created a new GUI interface for the smoke rack, this allows the play
to open the smoke rack GUI and look at the items hanging to determine if
they are dried or smoked without having to remove them from the smoke
rack.
2. Items can be added and removed from the smoke rack GUI and will
consume one Wool Yarn when they add an item.
3. Items can be still be added to the smoke rack without using the GUI,
by selected the item and clicking the smoke rack. This method still
consumes one Wool Yarn for each item added.
4. Items can be still be removed from the smoke rack without using the
GUI, by sneaking and clicking on the item to remove. The player must
have an empty hand to do this.
5. When clicking on the smoke rack and opening the GUI, if the player
does not have Wool Yarn in their inventory an error message is shown to
tell them.
6. After adding an item in the smoke rack GUI, if they run out of Wool
Yarn, the error message is shown they have no Wool Yarn.
7. If they try to add an item without using the GUI and they have no
Wool Yarn, the GUI will open instead of adding the item, and the error
message it shown.
8. Added an option to change the smoke rack material from Wool Yarn to
Jute Fibre.