Food cooking standaisation
Standardisaed how food is cooked; previously the forge would heat food
but not cook it so you couldn't eat it, the grill would cook it but not
impart any taste and the firepit worked. All code for cooking food has
been moved to TEFireEntity so it is cooked and taste imparted. This
should also mean that anyone building on TEFireEntity will inherit
correct food cooking and taste, unless they wish to override it.