Moved waila api to the libs folder and into it's own jar file,
that way gradle build will not inlude it as a part of the tfc package,
it will also be used as a dependency library in the build process. To be
used in eclipse it needs to be referenced as a Librery. (Maybe this
could be added to the "gradle eclipse")