Items melted in the forge that result in only 1 ingot will go into the input slot rather than try to be moved over to the storage slots. Bloom Overhaul - Crucible input slot will no longer accept refined blooms larger than 100 units, or raw blooms of any size. Raw blooms melted in the forge will give unknown ingots. Refined & Raw Blooms of any size melted in the forge will give at maximum 100 units of metal.
Note: This change means that all large blooms must be split in the anvil in order to get the full amount of wrought iron. Once the bloom has been split into 100 unit or less pieces, it is up to the player if they wish to turn it into an ingot by using the anvil, forge or crucible.