Feature: jug changes (this time without a ridiculous diff)
Adds the ability to define the number of notes in the range of the jug.

It will scale the notes based on the formula found here:
http://en.wikipedia.org/wiki/Piano_key_frequencies

This includes the fixes from my previous pull request.

My IDE decided to replace all the tabs with spaces last time :/

If you would like a pull of just the fixes please let me know.