Changed knapping interface to only consume the items from your inventory under specific conditions. Clay is consumed when the first valid recipe is formed. Rocks and leather are consumed when the first piece is removed from the screen.
Semi-fixed the duplication bug from using NEI to look up a recipe while in the knapping interface. Rocks and leather will properly save the state they were previously in when the player looks up a recipe. Clay will reset, but is still somewhat buggy. If the clay screen is reset before the output item is taken, the interface will look like it has been reset, but in reality all the same spots have still been removed (exit interface to truly reset without losing clay if there was no valid output). If the clay is reset after the output item is taken, the interface properly resets and you are able to form a new item which will consume another 5 clay.