Year length can no longer be set to a value less than 96. Cleaned up some logging messages. Entering invalid values in the config file will once again give notification that the value is incorrect in the console, and revert it to the default value. Removed the 1 Week button from the calendar debug screen. Added *Thumbs.db back to gitignore.
Existing worlds with shorter year lengths will have the value forced to be 96, which may result in some odd behavior. Most of this odd behavior can be fixed by taking note of the current in-game date before updating, and then using debug mode and the buttons on the calendar screen to advance time until it is that date again after updating. Players online while this is done will have their hunger and thirst bars drained, and all of the food in the world will experience at least some decay. Advance time using the largest increments possible first (years, then months, then days) to reduce the amount of food decay that happens.