New Registrations for Ore Dictionary Tools, and blocks.
The registration is primarily to allow for better integration with Hard
Core Questing (HQM) mod, but adds greater compatibility with other
Mods.

The order of some items listed in the 3 files was edited to be
consistent and clear.

The addition of wrought iron sheets in plateIron is for compatibility
as the TFC wrought iron is equivalent to iron in other mods.

- Registered all TFC tools such that there is registration of
ore:itemToolType for all and registration for material.

- Registration for all TFC weapons such that there is a registration
for ore:itemWeaponType for all and registration for material.

- Registration of vanilla and TFC Flint and Steel

- Registration of TFC chests in standard chestWood.

- Registration of TFC barrels in barrelWood.

- Registration of Vanilla and TFC Gravel in blockGravel.

- Registration of Seed Bages in bagSeed.

- Registration of wrought iron sheet in plateIron for compatibility.